feat: Add axios npm dependency for making HTTP requests
The code changes include adding the axios npm dependency to the package.json file. This dependency is used for making HTTP requests in the application. The axios version 1.7.3 is added to the dependencies list.

This commit message follows the established commit message conventions and provides a clear and concise description of the code changes.