feat: Add html-to-pdfmake helper for converting HTML to PDFMake format
The code changes include adding a new file, html-to-pdfmake.ts, which contains the getHtmlContent function. This function uses the html-to-pdfmake library and JSDOM to convert HTML content to PDFMake format. It is used to convert the HTML content of reports before generating PDFs. The html-to-pdfmake and jsdom dependencies are added to the package.json file.

This commit message follows the established commit message conventions and provides a clear and concise description of the code changes.